Provide preview buffer for test fixtures
This commit introduces a new preview buffer that displays the content of
the "Given:" block in an extra vertical window.
This is helpful for spotting specific line contents while editing test
verification code.

It provides new commands to open and close the preview window as well as
corresponding <Plug>-mappings.